Refactor comment commands to return DTOs, enhance authorization checks, and add extension methods for DTO conversion

+++ SchoolManagement.Application/Features/Comment/Command/DeleteComment/DeleteCommentCommand.cs
@@ -1,5 +1,7 @@
 using System.Net;
 using MediatR;
+using Microsoft.AspNetCore.Http;
+using SchoolManagement.Application.Services.AuthenticationService;
 using SchoolManagement.Application.Shared;
 using SchoolManagement.Domain.Interfaces.IRepositories;
 
@@ -13,9 +15,16 @@ public class DeleteCommentCommand : IRequest<Result<string>>
 public class DeleteCommentCommandHandler : IRequestHandler<DeleteCommentCommand, Result<string>>
 {
     private readonly ICommentRepositry _commentRepository;
-    public DeleteCommentCommandHandler(ICommentRepositry commentRepository)
+    private readonly IHttpContextAccessor _contextAccessor;
+    private readonly IUserAuthenticationService _authenticationService;
+    public DeleteCommentCommandHandler(
+        ICommentRepositry commentRepository,
+        IHttpContextAccessor contextAccessor,
+        IUserAuthenticationService authenticationService)
     {
         _commentRepository = commentRepository;
+        _contextAccessor = contextAccessor;
+        _authenticationService = authenticationService;
     }
     public async Task<Result<string>> Handle(DeleteCommentCommand request, CancellationToken cancellationToken)
     {
@@ -23,6 +32,11 @@ public class DeleteCommentCommandHandler : IRequestHandler<DeleteCommentCommand,
         if (comment == null)
         {
             return Result<string>.Failure("Comment does not exist.", HttpStatusCode.NotFound);
+        }
+        var user = await _authenticationService.GetCurrentUserAsync(_contextAccessor);
+        if (user.Id != comment.AuthorId)
+        {
+            return Result<string>.Failure("You are not authorized to delete this comment.", HttpStatusCode.Unauthorized);
         }
         await _commentRepository.Delete(comment);
         return Result<string>.SuccessMessage("Comment deleted successfully.");
